third_party/WebKit/LayoutTests/http/tests/navigation/beacon-cross-origin.https.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Verify navigator.sendBeacon() mixed content checking.");

window.jsTestIsAsync = true;

function test() {
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }

    shouldBeFalse('navigator.sendBeacon("http://localhost:8000/navigation/resources/save-beacon.php?name=cross-origin", "CrossOrigin");');
    finishJSTest();
}
</script>
</head>
<body onload="test();">
</body>
</html>
